Move lib tests from test/lib to test/unit/lib
