+++ test/unit/lib/unit/is-guest-checkout.js
@@ -1,7 +1,7 @@
 'use strict';
 
-var fake = require('../../helpers/fake');
-var isGuestCheckout = require('../../../src/lib/is-guest-checkout');
+var fake = require('../../../helpers/fake');
+var isGuestCheckout = require('../../../../src/lib/is-guest-checkout');
 
 describe('isGuestCheckout', function () {
   it('returns true when given a tokenization key', function () {
